@@ -15,6 +15,15 @@ fn parse_range(s: &str) -> Result<emver::VersionRange, JsValue> {
 }
 
 #[wasm_bindgen]
+pub fn flavor(version: &str) -> Result<JsValue, JsValue> {
+    parse_version(version).map(|v| {
+        v.flavor()
+            .map(wasm_bindgen::intern)
+            .map(JsValue::from_str)
+            .unwrap_or(JsValue::NULL)
+    })
+}
+#[wasm_bindgen]
 pub fn major(version: &str) -> Result<usize, JsValue> {
     parse_version(version).map(|v| v.major())
 }
@@ -31,14 +40,28 @@ pub fn revision(version: &str) -> Result<usize, JsValue> {
     parse_version(version).map(|v| v.revision())
 }
 #[wasm_bindgen]
-pub fn compare(lhs: &str, rhs: &str) -> Result<isize, JsValue> {
+pub fn prerelease(version: &str) -> Result<JsValue, JsValue> {
+    parse_version(version).map(|v| {
+        v.prerelease()
+            .as_deref()
+            .map(wasm_bindgen::intern)
+            .map(JsValue::from_str)
+            .unwrap_or(JsValue::NULL)
+    })
+}
+
+#[wasm_bindgen]
+pub fn compare(lhs: &str, rhs: &str) -> Result<JsValue, JsValue> {
     let s = parse_version(lhs).map_err(JsValue::from)?;
     let t = parse_version(rhs).map_err(JsValue::from)?;
-    Ok(match s.cmp(&t) {
-        Ordering::Less => -1,
-        Ordering::Equal => 0,
-        Ordering::Greater => 1,
-    })
+    Ok(s.partial_cmp(&t)
+        .map(|cmp| match cmp {
+            Ordering::Less => -1,
+            Ordering::Equal => 0,
+            Ordering::Greater => 1,
+        })
+        .map(JsValue::from)
+        .unwrap_or(JsValue::NULL))
 }
 #[wasm_bindgen]
 pub fn satisfies(version: &str, range: &str) -> Result<bool, JsValue> {
